consertando os bugs das reservas
