+++ src/app/(admin)/profile.tsx
@@ -1,0 +1,99 @@
+import React from 'react';
+import { View, Text, StyleSheet, TouchableOpacity, Alert } from 'react-native';
+import { useRouter } from 'expo-router';
+import { auth } from '../../config/firebaseConfig';
+import { Ionicons } from '@expo/vector-icons';
+import { SafeAreaView } from 'react-native-safe-area-context';
+
+const HEADER = '#5145CD';
+const WHITE = '#FFFFFF';
+
+export default function AdminProfileScreen() {
+  const router = useRouter();
+
+  const handleLogout = async () => {
+    try {
+      await auth.signOut();
+      router.replace('/login');
+    } catch (error) {
+      Alert.alert('Erro', 'Não foi possível sair no momento.');
+    }
+  };
+
+  return (
+    <SafeAreaView edges={['top']} style={styles.container}>
+      <View style={styles.header}>
+        <Text style={styles.title}>Perfil Admin</Text>
+      </View>
+
+      <View style={styles.content}>
+        <View style={styles.avatarCircle}>
+          <Ionicons name="person" size={40} color={HEADER} />
+        </View>
+        <Text style={styles.emailText}>{auth.currentUser?.email || 'Admin'}</Text>
+
+        <TouchableOpacity 
+          style={styles.logoutButton} 
+          onPress={handleLogout}
+          activeOpacity={0.8}
+        >
+          <Ionicons name="log-out-outline" size={24} color={WHITE} style={{ marginRight: 8 }} />
+          <Text style={styles.logoutText}>Sair da Conta</Text>
+        </TouchableOpacity>
+      </View>
+    </SafeAreaView>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#F4F5F8',
+  },
+  header: {
+    backgroundColor: HEADER,
+    padding: 24,
+    alignItems: 'center',
+    borderBottomLeftRadius: 20,
+    borderBottomRightRadius: 20,
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    color: WHITE,
+  },
+  content: {
+    flex: 1,
+    alignItems: 'center',
+    padding: 24,
+    marginTop: 30,
+  },
+  avatarCircle: {
+    width: 100,
+    height: 100,
+    borderRadius: 50,
+    backgroundColor: '#E0E7FF',
+    justifyContent: 'center',
+    alignItems: 'center',
+    marginBottom: 16,
+  },
+  emailText: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    color: '#1F2937',
+    marginBottom: 40,
+  },
+  logoutButton: {
+    backgroundColor: '#EF4444',
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingVertical: 14,
+    paddingHorizontal: 32,
+    borderRadius: 12,
+  },
+  logoutText: {
+    color: WHITE,
+    fontSize: 16,
+    fontWeight: 'bold',
+  }
+});
