+++ src/app/index.tsx
@@ -36,7 +36,7 @@ export default function LandingScreen() {
           const role = userDoc.exists() ? userDoc.data()?.role : 'user';
 
           if (role === 'admin') {
-            router.replace('/dashboard');
+            router.replace('/(admin)');
           } else {
             router.replace('/(tabs)/kits_list');
           }
